#1 changed project structure
