--- valve/valve_ds/__init__.py
@@ -1,6 +1,0 @@
-"""
-This package contains Valve Device Class based on facadedevice library
-"""
-
-__author__ = "Stanislaw Cabala"
-__author_email__ = "stanislaw.cabala@gmail.com"
